refactor: domain Enum 클래스 하위 패키지로 위치 이동
