+++ src/main/kotlin/org/lgm/jobboard/jobposting/infra/persistence/JobPostingRepository.kt
@@ -1,6 +1,6 @@
 package org.lgm.jobboard.jobposting.infra.persistence
 
-import org.lgm.jobboard.jobposting.domain.JobPostingStatus
+import org.lgm.jobboard.jobposting.domain.type.JobPostingStatus
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
