+++ src/main/kotlin/org/lgm/jobboard/outbox/application/port/OutboxCommandPort.kt
@@ -1,7 +1,7 @@
 package org.lgm.jobboard.outbox.application.port
 
-import org.lgm.jobboard.outbox.domain.AggregateType
-import org.lgm.jobboard.outbox.domain.EventType
+import org.lgm.jobboard.outbox.domain.type.AggregateType
+import org.lgm.jobboard.outbox.domain.type.EventType
 
 interface OutboxCommandPort {
 	fun enqueue(
