--- src/main/kotlin/org/lgm/jobboard/outbox/domain/AggregateType.kt
@@ -1,5 +1,0 @@
-package org.lgm.jobboard.outbox.domain
-
-enum class AggregateType {
-	JOB_POSTING
-}
