--- src/main/kotlin/org/lgm/jobboard/outbox/domain/EventType.kt
@@ -1,6 +1,0 @@
-package org.lgm.jobboard.outbox.domain
-
-enum class EventType {
-	UPSERT,
-	DELETE
-}
