+++ src/main/kotlin/org/lgm/jobboard/outbox/domain/type/AggregateType.kt
@@ -1,0 +1,5 @@
+package org.lgm.jobboard.outbox.domain.type
+
+enum class AggregateType {
+	JOB_POSTING
+}
