+++ src/main/kotlin/org/lgm/jobboard/outbox/domain/type/EventType.kt
@@ -1,0 +1,6 @@
+package org.lgm.jobboard.outbox.domain.type
+
+enum class EventType {
+	UPSERT,
+	DELETE
+}
