+++ src/main/kotlin/org/lgm/jobboard/outbox/infra/persistence/adapter/OutboxCommandAdapter.kt
@@ -1,8 +1,8 @@
 package org.lgm.jobboard.outbox.infra.persistence.adapter
 
 import org.lgm.jobboard.outbox.application.port.OutboxCommandPort
-import org.lgm.jobboard.outbox.domain.AggregateType
-import org.lgm.jobboard.outbox.domain.EventType
+import org.lgm.jobboard.outbox.domain.type.AggregateType
+import org.lgm.jobboard.outbox.domain.type.EventType
 import org.lgm.jobboard.outbox.infra.persistence.OutboxEventEntity
 import org.lgm.jobboard.outbox.infra.persistence.OutboxEventRepository
 import org.springframework.stereotype.Component
